+++ cnd/src/herc20.rs
@@ -88,8 +88,7 @@ where
             )
             .await
             {
-                Ok(comit::herc20::Funded::Correctly { transaction, asset }) => {
-                    let event = comit::swap::herc20::Funded { transaction, asset };
+                Ok(Ok(event)) => {
                     self.storage
                         .herc20_events
                         .lock()
@@ -100,12 +99,7 @@ where
 
                     return Ok(event);
                 }
-                Ok(comit::herc20::Funded::Incorrectly { asset, .. }) => {
-                    return Err(IncorrectlyFunded {
-                        expected: params.asset,
-                        got: asset,
-                    })
-                }
+                Ok(Err(e)) => return Err(e),
                 Err(e) => {
                     tracing::warn!("failed to watch for herc20 funding, retrying ...: {:#}", e)
                 }
