+++ comit/src/hbit.rs
@@ -19,19 +19,21 @@ use bitcoin::{
 };
 use blockchain_contracts::bitcoin::{hbit::Htlc, witness::UnlockParameters};
 use std::cmp::Ordering;
+use thiserror::Error;
 use time::OffsetDateTime;
 use tracing_futures::Instrument;
 
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum Funded {
-    Correctly {
-        asset: asset::Bitcoin,
-        location: htlc_location::Bitcoin,
-    },
-    Incorrectly {
-        asset: asset::Bitcoin,
-        location: htlc_location::Bitcoin,
-    },
+#[derive(Debug, Clone, Copy, Error)]
+#[error("hbit HTLC was incorrectly funded, expected {expected} but got {got}")]
+pub struct IncorrectlyFunded {
+    pub expected: asset::Bitcoin,
+    pub got: asset::Bitcoin,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct Funded {
+    pub asset: asset::Bitcoin,
+    pub location: htlc_location::Bitcoin,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -45,11 +47,30 @@ pub struct Refunded {
     pub transaction: bitcoin::Txid,
 }
 
+#[async_trait::async_trait]
+pub trait WatchForFunded {
+    async fn watch_for_funded(
+        &self,
+        params: &Params,
+        start_of_swap: OffsetDateTime,
+    ) -> Result<Funded, IncorrectlyFunded>;
+}
+
+#[async_trait::async_trait]
+pub trait WatchForRedeemed {
+    async fn watch_for_redeemed(
+        &self,
+        params: &Params,
+        fund_event: Funded,
+        start_of_swap: OffsetDateTime,
+    ) -> Redeemed;
+}
+
 pub async fn watch_for_funded<C>(
     connector: &C,
-    params: &Params,
+    params: &SharedParams,
     start_of_swap: OffsetDateTime,
-) -> Result<Funded>
+) -> Result<Result<Funded, IncorrectlyFunded>>
 where
     C: LatestBlock<Block = Block>
         + BlockByHash<Block = Block, BlockHash = BlockHash>
@@ -64,17 +85,18 @@ where
 
     let asset = asset::Bitcoin::from_sat(transaction.output[location.vout as usize].value);
 
-    let event = match expected_asset.cmp(&asset) {
-        Ordering::Equal => Funded::Correctly { asset, location },
-        _ => Funded::Incorrectly { asset, location },
-    };
-
-    Ok(event)
+    match expected_asset.cmp(&asset) {
+        Ordering::Equal => Ok(Ok(Funded { asset, location })),
+        _ => Ok(Err(IncorrectlyFunded {
+            expected: expected_asset,
+            got: asset,
+        })),
+    }
 }
 
 pub async fn watch_for_redeemed<C>(
     connector: &C,
-    params: &Params,
+    params: &SharedParams,
     location: htlc_location::Bitcoin,
     start_of_swap: OffsetDateTime,
 ) -> Result<Redeemed>
@@ -99,7 +121,7 @@ where
 
 pub async fn watch_for_refunded<C>(
     connector: &C,
-    params: &Params,
+    params: &SharedParams,
     location: htlc_location::Bitcoin,
     start_of_swap: OffsetDateTime,
 ) -> Result<Refunded>
@@ -119,7 +141,7 @@ where
 }
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct Params {
+pub struct SharedParams {
     pub network: ledger::Bitcoin,
     pub asset: asset::Bitcoin,
     pub redeem_identity: identity::Bitcoin,
@@ -128,7 +150,14 @@ pub struct Params {
     pub secret_hash: SecretHash,
 }
 
-impl Params {
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct Params {
+    pub shared: SharedParams,
+    pub transient_sk: SecretKey,
+    pub final_address: bitcoin::Address,
+}
+
+impl SharedParams {
     pub fn build_fund_action(&self) -> SendToAddress {
         let network = self.network;
         let to = self.compute_address();
@@ -223,8 +252,8 @@ impl Params {
     }
 }
 
-impl From<Params> for Htlc {
-    fn from(params: Params) -> Self {
+impl From<SharedParams> for Htlc {
+    fn from(params: SharedParams) -> Self {
         build_bitcoin_htlc(
             params.redeem_identity,
             params.refund_identity,
@@ -234,7 +263,7 @@ impl From<Params> for Htlc {
     }
 }
 
-impl Params {
+impl SharedParams {
     pub fn compute_address(&self) -> Address {
         Htlc::from(*self).compute_address(self.network.into())
     }
@@ -270,6 +299,63 @@ pub fn build_bitcoin_htlc(
         redeem_identity,
         secret_hash.into_raw(),
     )
+}
+
+#[cfg(feature = "quickcheck")]
+mod arbitrary {
+    use super::*;
+    use crate::{asset, identity, ledger, SecretHash, Timestamp};
+    use ::bitcoin::secp256k1::SecretKey;
+    use quickcheck::{Arbitrary, Gen};
+
+    impl Arbitrary for Params {
+        fn arbitrary<G: Gen>(g: &mut G) -> Self {
+            Params {
+                shared: SharedParams {
+                    network: bitcoin_network(g),
+                    asset: bitcoin_asset(g),
+                    redeem_identity: bitcoin_identity(g),
+                    refund_identity: bitcoin_identity(g),
+                    expiry: Timestamp::arbitrary(g),
+                    secret_hash: SecretHash::arbitrary(g),
+                },
+                transient_sk: secret_key(g),
+                final_address: bitcoin_address(g),
+            }
+        }
+    }
+
+    fn secret_key<G: Gen>(g: &mut G) -> SecretKey {
+        let mut bytes = [0u8; 32];
+        for byte in &mut bytes {
+            *byte = u8::arbitrary(g);
+        }
+        SecretKey::from_slice(&bytes).unwrap()
+    }
+
+    fn bitcoin_network<G: Gen>(g: &mut G) -> ledger::Bitcoin {
+        match u8::arbitrary(g) % 3 {
+            0 => ledger::Bitcoin::Mainnet,
+            1 => ledger::Bitcoin::Testnet,
+            2 => ledger::Bitcoin::Regtest,
+            _ => unreachable!(),
+        }
+    }
+
+    fn bitcoin_asset<G: Gen>(g: &mut G) -> asset::Bitcoin {
+        asset::Bitcoin::from_sat(u64::arbitrary(g))
+    }
+
+    fn bitcoin_identity<G: Gen>(g: &mut G) -> identity::Bitcoin {
+        identity::Bitcoin::from_secret_key(
+            &bitcoin::secp256k1::Secp256k1::signing_only(),
+            &secret_key(g),
+        )
+    }
+
+    fn bitcoin_address<G: Gen>(g: &mut G) -> bitcoin::Address {
+        bitcoin::Address::p2wpkh(&bitcoin_identity(g).into(), bitcoin_network(g).into()).unwrap()
+    }
 }
 
 #[cfg(test)]
